requirements.txt
flask==2.2.5
markupsafe==2.1.3
gunicorn==20.1.0
pillow==11.3.0


